Reworked mappers
