+++ src/UserService.Business/CreateUserCommand.cs
@@ -2,8 +2,7 @@
 using LT.DigitalOffice.Kernel.FluentValidationExtensions;
 using LT.DigitalOffice.UserService.Business.Interfaces;
 using LT.DigitalOffice.UserService.Data.Interfaces;
-using LT.DigitalOffice.UserService.Mappers.Interfaces;
-using LT.DigitalOffice.UserService.Models.Db;
+using LT.DigitalOffice.UserService.Mappers.RequestsMappers.Interfaces;
 using LT.DigitalOffice.UserService.Models.Dto;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -15,14 +14,14 @@ namespace LT.DigitalOffice.UserService.Business
     {
         private readonly IUserRepository _userRepository;
         private readonly IValidator<UserRequest> _validator;
-        private readonly IMapper<UserRequest, DbUser> _mapperUser;
-        private readonly IMapper<UserRequest, DbUserCredentials> _mapperUserCredentials;
+        private readonly IUserRequestMapper _mapperUser;
+        private readonly IUserCredentialsRequestMapper _mapperUserCredentials;
 
         public CreateUserCommand(
             [FromServices] IUserRepository userRepository,
             [FromServices] IValidator<UserRequest> validator,
-            [FromServices] IMapper<UserRequest, DbUser> mapperUser,
-            [FromServices] IMapper<UserRequest, DbUserCredentials> mapperUserCredentials)
+            [FromServices] IUserRequestMapper mapperUser,
+            [FromServices] IUserCredentialsRequestMapper mapperUserCredentials)
         {
             _validator = validator;
             _userRepository = userRepository;
