+++ src/UserService.Business/EditUserCommand.cs
@@ -2,27 +2,27 @@
 using LT.DigitalOffice.Kernel.FluentValidationExtensions;
 using LT.DigitalOffice.UserService.Business.Interfaces;
 using LT.DigitalOffice.UserService.Data.Interfaces;
-using LT.DigitalOffice.UserService.Mappers.Interfaces;
-using LT.DigitalOffice.UserService.Models.Db;
+using LT.DigitalOffice.UserService.Mappers.RequestsMappers.Interfaces;
 using LT.DigitalOffice.UserService.Models.Dto;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LT.DigitalOffice.UserService.Business
 {
+    /// <inheritdoc/>
     public class EditUserCommand : IEditUserCommand
     {
         private readonly IValidator<UserRequest> validator;
         private readonly IUserRepository userRepository;
         private readonly IUserCredentialsRepository userCredentialsRepository;
-        private readonly IMapper<UserRequest, DbUser> mapperUser;
-        private readonly IMapper<UserRequest, DbUserCredentials> mapperUserCredentials;
+        private readonly IUserRequestMapper mapperUser;
+        private readonly IUserCredentialsRequestMapper mapperUserCredentials;
 
         public EditUserCommand(
             [FromServices] IValidator<UserRequest> validator,
             [FromServices] IUserRepository userRepository,
             [FromServices] IUserCredentialsRepository userCredentialsRepository,
-            [FromServices] IMapper<UserRequest, DbUser> mapperUser,
-            [FromServices] IMapper<UserRequest, DbUserCredentials> mapperUserCredentials)
+            [FromServices] IUserRequestMapper mapperUser,
+            [FromServices] IUserCredentialsRequestMapper mapperUserCredentials)
         {
             this.validator = validator;
             this.userRepository = userRepository;
@@ -31,6 +31,7 @@ namespace LT.DigitalOffice.UserService.Business
             this.userCredentialsRepository = userCredentialsRepository;
         }
 
+        /// <inheritdoc/>
         public bool Execute(UserRequest request)
         {
             validator.ValidateAndThrowCustom(request);
