+++ src/UserService.Business/GetUserByEmailCommand.cs
@@ -2,29 +2,30 @@
 using LT.DigitalOffice.Kernel.FluentValidationExtensions;
 using LT.DigitalOffice.UserService.Business.Interfaces;
 using LT.DigitalOffice.UserService.Data.Interfaces;
-using LT.DigitalOffice.UserService.Mappers.Interfaces;
-using LT.DigitalOffice.UserService.Models.Db;
+using LT.DigitalOffice.UserService.Mappers.ResponsesMappers.Interfaces;
 using LT.DigitalOffice.UserService.Models.Dto;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LT.DigitalOffice.UserService.Business
 {
+    /// <inheritdoc/>
     public class GetUserByEmailCommand : IGetUserByEmailCommand
     {
         private readonly IValidator<string> validator;
         private readonly IUserRepository repository;
-        private readonly IMapper<DbUser, User> mapper;
+        private readonly IUserResponseMapper mapper;
 
         public GetUserByEmailCommand(
             [FromServices] IValidator<string> validator,
             [FromServices] IUserRepository repository,
-            [FromServices] IMapper<DbUser, User> mapper)
+            [FromServices] IUserResponseMapper mapper)
         {
             this.validator = validator;
             this.repository = repository;
             this.mapper = mapper;
         }
 
+        /// <inheritdoc/>
         public User Execute(string userEmail)
         {
             validator.ValidateAndThrowCustom(userEmail);
